Reworked recursive iteration through blocks to be more elegant
